File Name and Structure clean up. Add in Desconstructor for Test Harness.
File Name and Structure clean up. Add in Desconstructor for Test Harness.
