--- DWardProject1/DWardProject1.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.28803.202
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DWardProject1", "DWardProject1.vcxproj", "{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Debug|x64.ActiveCfg = Debug|x64
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Debug|x64.Build.0 = Debug|x64
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Debug|x86.ActiveCfg = Debug|Win32
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Debug|x86.Build.0 = Debug|Win32
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Release|x64.ActiveCfg = Release|x64
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Release|x64.Build.0 = Release|x64
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Release|x86.ActiveCfg = Release|Win32
-		{45FFD196-47D0-4A9B-AEB5-AE459F2ED850}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {50D1BFD4-912A-4CBF-B210-AA10519BA63F}
-	EndGlobalSection
-EndGlobal
