--- DWardProject1/Debug/DWardProject1.Build.CppClean.log
@@ -1,15 +1,0 @@
-c:\users\derrick\documents\c++\dwardproject1\debug\vc142.pdb
-c:\users\derrick\documents\c++\dwardproject1\debug\vc142.idb
-c:\users\derrick\documents\c++\dwardproject1\debug\testpredicate.obj
-c:\users\derrick\documents\c++\dwardproject1\debug\testharness.obj
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.obj
-c:\users\derrick\documents\c++\dwardproject1\debug\demounittests.obj
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.exe
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.ilk
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.pdb
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.tlog\cl.command.1.tlog
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.tlog\cl.read.1.tlog
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.tlog\cl.write.1.tlog
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.tlog\link.command.1.tlog
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.tlog\link.read.1.tlog
-c:\users\derrick\documents\c++\dwardproject1\debug\dwardproject1.tlog\link.write.1.tlog
