--- DWardProject1/Debug/DWardProject1.log
@@ -1,6 +1,0 @@
-﻿  DemoUnitTests.cpp
-  DWardProject1.cpp
-  TestHarness.cpp
-  TestPredicate.cpp
-  Generating Code...
-  DWardProject1.vcxproj -> C:\Users\Derrick\Documents\C++\DWardProject1\Debug\DWardProject1.exe
