+++ Debug/Project1.log
@@ -1,0 +1,5 @@
+﻿  Project1.cpp
+  TestHarness.cpp
+  UnitTests.cpp
+  Generating Code...
+  Project1.vcxproj -> C:\Users\dewa\Documents\GitHub\CSE-687-project-1\Debug\Project1.exe
