+++ Project1.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.28803.452
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Project1", "Project1.vcxproj", "{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Debug|x64.ActiveCfg = Debug|x64
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Debug|x64.Build.0 = Debug|x64
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Debug|x86.ActiveCfg = Debug|Win32
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Debug|x86.Build.0 = Debug|Win32
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Release|x64.ActiveCfg = Release|x64
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Release|x64.Build.0 = Release|x64
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Release|x86.ActiveCfg = Release|Win32
+		{1CE13BF4-FC86-4091-8E6D-E8D8CCCA79E8}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {5FC6FB5F-CE8C-4B75-9565-E629FA969450}
+	EndGlobalSection
+EndGlobal
