+++ UnitTests.h
@@ -1,0 +1,18 @@
+/*
+Name: Derrick Ward
+Class/Section: CSE 687
+Project: 1
+Description: This file is the interface for a listing of unit tests
+Date: 10/15/2019
+*/
+
+#include "TestPredicate.h"
+
+#ifndef UNITTESTS
+#define UNITTESTS
+
+TestPredicate unitTest1();
+TestPredicate unitTest2();
+TestPredicate unitTest3();
+
+#endif
